ci: don't run tests on overlayfs/tmpfs
It occurred to me that a simple way of getting coverage for tests that
don't work on overlayfs/tmpfs is to simply run them in a directory
mounted from the host.

To do this, we switch `f26-primary` to execute on a test node rather
than in a container, and make use of the new `TEST_TMPDIR` var to point
it at the mounted /srv dir.

We also have to tweak the overlayfs detection logic only check the
filesystem in which we're actually running the test, rather than just /.